requirements.txt
streamlit>=1.22.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
nltk>=3.8.0
python-docx>=0.8.0
PyPDF2>=3.0.0
plotly>=5.14.0
wordcloud>=1.9.0
pywin32>=300; sys_platform == 'win32'